agent/requirements.txt
google-adk
google-cloud-aiplatform
kagglehub
pandas
fastapi
uvicorn
